Resolución de tipos

+++ Assets/Scripts/NexusBehaviour.cs
@@ -38,7 +38,7 @@ public class NexusBehaviour : NetworkBehaviour {
 			if (unit.isBoss)
 				playerId.TakeDamage ((int)((float) unit.health * unit.damage / (float)unit.max_health));
 			else
-				playerId.TakeDamage (unit.damage);
+				playerId.TakeDamage ((int)unit.damage);
 			
 			syncOwner.getOwner ().GetComponent<PlayerId> ().GainMoney (unit.money);
 			Destroy(other.gameObject);
